Make compile with -Wall.  This included adding protos.h to the talker directory, to contain a few prototypes that are needed (possibly more later).  Also includes casting args to the ctype.h is* macros.

+++ src/spider/modules/message/main.c
@@ -136,7 +136,7 @@ main (void)
     {
       getcmd (buf);
       /* A command? */
-      if (!isdigit (buf[0]))
+      if (!isdigit ((int)buf[0]))
 	{
 	  if (parse (&cmdline, buf))
 	    /* We have something to reply to */
